missing cast

+++ array.h
@@ -18,7 +18,7 @@
 	ssize_t __end = (end);					\
 	CHECK(__start >= 0);					\
 	CHECK(__end >= __start);				\
-	CHECK(__end <= array_lengthof(*__x));			\
+	CHECK((size_t)__end <= array_lengthof(*__x));		\
 	(array(typeof((*__x)[0]), __end - __start)*)&(*__x)[__start];	\
 }))
